+++ app/src/main/java/com/app/arcabyolimpo/data/remote/api/ArcaApi.kt
@@ -13,6 +13,7 @@ import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordDto
 import com.app.arcabyolimpo.data.remote.dto.password.UpdatePasswordResponseDto
 import com.app.arcabyolimpo.data.remote.dto.password.VerifyTokenResponseDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.AcquisitionDto
+import com.app.arcabyolimpo.data.remote.dto.productbatches.ProductBatchDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.DeleteResponseDto
 import com.app.arcabyolimpo.data.remote.dto.supplies.GetFiltersDto
@@ -138,8 +139,14 @@ interface ArcaApi {
     ): DeleteResponseDto
 
     // Workshop ---------------------------
+
     @GET("workshop")
     suspend fun getWorkshopsList(): List<WorkshopsListDto>
+
+    @GET("workshop/search")
+    suspend fun searchWorkshops(
+        @Query("nameWorkshop") name: String
+    ): List<WorkshopDto>
 
     @GET("workshop/{id}")
     suspend fun getWorkshop(
@@ -152,7 +159,7 @@ interface ArcaApi {
     ): AddNewWorkshopDto
 
     // Beneficiary -------------
-    @GET("beneficiary")
+    @GET("beneficiary/list")
     suspend fun getBeneficiariesList(): List<BeneficiariesListDto>
 
     @GET("beneficiary/{id}")
@@ -171,11 +178,36 @@ interface ArcaApi {
     @Multipart
     @POST("supplies/add")
     suspend fun addSupply(
-        @Part("idWorkshop") idWorkshop: RequestBody,
-        @Part("name") name: RequestBody,
-        @Part("measureUnit") measureUnit: RequestBody,
-        @Part("idCategory") idCategory: RequestBody,
+        @Part("idTaller") idWorkshop: RequestBody,
+        @Part("nombre") name: RequestBody,
+        @Part("unidadMedida") measureUnit: RequestBody,
+        @Part("idCategoria") idCategory: RequestBody,
         @Part("status") status: RequestBody,
-        @Part image: MultipartBody.Part?,
+        @Part imagenInsumo: MultipartBody.Part?
+    )
+
+    @GET("productBatch/")
+    suspend fun getProductBatches(): List<ProductBatchDto>
+
+    @GET("productBatch/{id}")
+    suspend fun getProductBatch(
+        @Path("id") id: String,
+    ): ProductBatchDto
+
+    @POST("productBatch/")
+    suspend fun addProductBatch(
+        @Body batch: ProductBatchDto,
+    )
+
+    @Multipart
+    @POST("product/add")
+    suspend fun addProduct(
+        @Part("idTaller") idWorkshop: RequestBody,
+        @Part("Nombre") name: RequestBody,
+        @Part("PrecioUnitario") unitaryPrice: RequestBody,
+        @Part("idCategoria") idCategory: RequestBody,
+        @Part("Descripcion") description: RequestBody,
+        @Part("Disponible") status: RequestBody,
+        @Part image: MultipartBody.Part?
     )
 }