+++ app/src/main/java/com/app/arcabyolimpo/data/repository/beneficiaries/BeneficiaryRepositoryImpl.kt
@@ -35,16 +35,22 @@ class BeneficiaryRepositoryImpl
         override suspend fun getBeneficiariesList(): List<Beneficiary> {
             val response = api.getBeneficiariesList()
             return response.map { dto ->
+                val fullName = listOfNotNull(
+                    dto.firstName,
+                    dto.paternalName,
+                    dto.maternalName
+                ).joinToString(" ").trim()
+
                 Beneficiary(
-                    id = dto.id,
-                    name = dto.name,
+                    id = dto.id.orEmpty(),
+                    name = fullName.ifEmpty{ "Nombre no disponible" },
                     birthdate = "",
                     emergencyNumber = "",
                     emergencyName = "",
                     emergencyRelation = "",
                     details = "",
                     entryDate = "",
-                    image = dto.image,
+                    image = dto.image.orEmpty(),
                     disabilities = "",
                     status = 0,
                 )