+++ app/src/main/java/com/app/arcabyolimpo/di/AppModule.kt
@@ -9,15 +9,20 @@ import com.app.arcabyolimpo.data.remote.interceptor.TokenAuthenticator
 import com.app.arcabyolimpo.data.repository.auth.UserRepositoryImpl
 import com.app.arcabyolimpo.data.repository.beneficiaries.BeneficiaryRepositoryImpl
 import com.app.arcabyolimpo.data.repository.password.PasswordUserRepositoryImpl
+import com.app.arcabyolimpo.data.repository.productbatches.ProductBatchRepositoryImpl
+import com.app.arcabyolimpo.data.repository.product.ProductRepositoryImpl
 import com.app.arcabyolimpo.data.repository.supplies.SupplyRepositoryImpl
 import com.app.arcabyolimpo.data.repository.user.UsersRepositoryImpl
 import com.app.arcabyolimpo.data.repository.workshops.WorkshopRepositoryImpl
 import com.app.arcabyolimpo.domain.repository.auth.UserRepository
 import com.app.arcabyolimpo.domain.repository.beneficiaries.BeneficiaryRepository
 import com.app.arcabyolimpo.domain.repository.password.PasswordUserRepository
+import com.app.arcabyolimpo.domain.repository.productbatches.ProductBatchRepository
+import com.app.arcabyolimpo.domain.repository.product.ProductRepository
 import com.app.arcabyolimpo.domain.repository.supplies.SupplyRepository
 import com.app.arcabyolimpo.domain.repository.user.UsersRepository
 import com.app.arcabyolimpo.domain.repository.workshops.WorkshopRepository
+import com.app.arcabyolimpo.domain.usecase.productbatches.GetProductBatchesUseCase
 import dagger.Module
 import dagger.Provides
 import dagger.hilt.InstallIn
@@ -32,7 +37,7 @@ import javax.inject.Singleton
 @Module
 @InstallIn(SingletonComponent::class)
 object AppModule {
-    private const val BASE_URL = "https://magen-nonsectional-rosette.ngrok-free.dev/" // LOCALHOST
+    private const val BASE_URL = "http://10.0.2.2:8080/" // LOCALHOST
 
     /**
      * Provides a configured [OkHttpClient] instance.
@@ -119,6 +124,19 @@ object AppModule {
     ): SupplyRepository = SupplyRepositoryImpl(api, context)
 
     /**
+     * Provides the [ProductRepository] implementation.
+     *
+     * @param api The [ArcaApi] instance used to perform network requests.
+     * @return A singleton instance of [ProductRepositoryImpl].
+     */
+    @Provides
+    @Singleton
+    fun provideProductRepository(
+        api: ArcaApi,
+        @ApplicationContext context: Context,
+    ): ProductRepository = ProductRepositoryImpl(api, context)
+
+    /**
      * Provides the [WorkshopRepository] implementation.
      *
      * This repository handles all workshop-related data operations,
@@ -151,4 +169,9 @@ object AppModule {
     @Provides
     @Singleton
     fun provideUsersRepository(api: ArcaApi): UsersRepository = UsersRepositoryImpl(api)
+
+    /** Provides the [ProductBatchRepository] implementation.*/
+    @Provides
+    @Singleton
+    fun provideProductBatchRepository(api: ArcaApi): ProductBatchRepository = ProductBatchRepositoryImpl(api)
 }