+++ app/src/main/java/com/app/arcabyolimpo/domain/repository/workshops/WorkshopRepository.kt
@@ -11,4 +11,8 @@ interface WorkshopRepository {
 
     suspend fun addWorkshop(newWorkshop: WorkshopDto): Workshop
 
+    suspend fun searchWorkshop(name: String): List<Workshop>
+
+    companion object
+
 }